Moving the AbuseIPDB Widget to an observer class
As the title implies, this moves the display of the AbuseIPDB contributor badge from a self-install "optional" steps method to an optional enable by making it a ZenCart observer class.

This starts 3.1.0, but I'm leaving it up to the main maintainer to decide whether or not to change the version number.