Moving the widget to an observer class
- Changed the directions of the installation to match
- Reformatted README.md to be within spec for Markdown
- Adding directions to remove the old AbuseIPDBDashboardWidget.php if it wasn't already deleted.
- Changed the optional widget to an observer class instead of manual hard edits. (This way it's not accidentally overwritten on a ZenCart upgrade.)

(Reissue of issue #11 and #10, stuff got lost in transitioning to the an observer.)

The problem with this file is that it was still trying to find the widget in the original DIR_WS_MODULES directory which made no sense since it's not there. Instead of looking there, the observer will not look to the zc_plugins directory as it should be.

Since there is presently no "neat" way of doing this, the observer will extract the current version of the module, as defined in the database, and look for the AbuseIPDBDashboardWidget.php file in THAT module's admin/includes/modules/dashboard_widgets directory.